Add files via upload
This project uses Keras to create an LSTM attention model in order to translate various date formats e.g. ( "the 29th of August 1958", "03/30/1968", "24 JUNE 1987") to machine readable dates (e.g. "1958-08-29", "1968-03-30", "1987-06-24").  The attention model is ideal for this task because the first attention layer of LSTMs decides which context is important for the 2nd layer of LSTMs to generate each standardized date digit.