Trigger word detection - v1
For trigger word detection, like with "okay google" or "hey Siri", an LSTM network can be used.  The audio file can be descritized into time slots with frequency domain amplitudes.  Then this frequency feature can be used as a time series with LSTM to generate a series of 0's and 1's, where 1's indicate the trigger word has been said.